move NavHeader to top level
