+++ app/components/navheader/index.js
@@ -7,7 +7,7 @@ import DemographicInfo from './DemographicInfo';
 import PatientMenuOptions from './MenuOptions/Patient';
 import ClinicianMenuOptions from './MenuOptions/Clinician';
 
-const UploadLaunchOverlay = require('../../../components/uploadlaunchoverlay');
+const UploadLaunchOverlay = require('../../components/uploadlaunchoverlay');
 
 const innerContainerStyleProps = {
   px: 4, 
@@ -21,11 +21,11 @@ const innerContainerStyleProps = {
   }
 };
 
-const PatientDataHeader = ({ patient, isUserPatient }) => {
+const NavHeader = ({ patient, isUserPatient }) => {
   const history = useHistory();
   const [isUploadOverlayOpen, setIsUploadOverlayOpen] = useState(false);
 
-  if (!patient.profile) return null; // not available immediately on component mount
+  if (!patient?.profile) return null; // not available immediately on component mount
 
   const handleUpload      = () => setIsUploadOverlayOpen(true);
   const handleViewData    = () => history.push(`/patients/${patient.userid}/data`);
@@ -63,4 +63,4 @@ const PatientDataHeader = ({ patient, isUserPatient }) => {
   );
 }
 
-export default PatientDataHeader;
+export default NavHeader;
