+++ app/pages/patient/patient.js
@@ -19,7 +19,6 @@ import React from 'react';
 import createReactClass from 'create-react-class';
 import { withTranslation, Trans } from 'react-i18next';
 import _ from 'lodash';
-import PatientDataHeader from '../patientdata/PatientDataHeader';
 
 import personUtils from '../../core/personutils';
 import ModalOverlay from '../../components/modaloverlay';
@@ -73,19 +72,13 @@ const Patient = withTranslation()(createReactClass({
 
   render: function() {
     return (
-      <>
-        <PatientDataHeader 
-          patient={this.props.patient}
-          isUserPatient={this.isSamePersonUserAndPatient()} 
-        />
-        <div className="PatientPage js-patient-page">
-          <div className="PatientPage-layer">
-            {this.renderSubnav()}
-            {this.renderContent()}
-            {this.renderFooter()}
-          </div>
+      <div className="PatientPage js-patient-page">
+        <div className="PatientPage-layer">
+          {this.renderSubnav()}
+          {this.renderContent()}
+          {this.renderFooter()}
         </div>
-      </>
+      </div>
     );
   },
